python_scripts/pyproject.toml
[project]
name = "python-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-firestore>=2.20.0",
    "pydantic>=2.10.5",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.0",
]
